Admin page 1/2, admin middleware, enhanced AwardsService

+++ database/migrations/2021_06_15_011553_add_coins_to_users_table.php
@@ -14,7 +14,7 @@ class AddCoinsToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->mediumInteger('coins')->default('0');
+            $table->bigInteger('coins')->default('0');
         });
     }
